+++ packages/server/src/adapters/web.ts
@@ -2,7 +2,8 @@
  * Web platform adapter implementing IPlatformAdapter with SSE stream management.
  * Bridge between the orchestrator and the React frontend via Server-Sent Events.
  */
-import type { IWebPlatformAdapter, MessageChunk, MessageMetadata } from '@archon/core';
+import type { IWebPlatformAdapter, MessageMetadata } from '@archon/core';
+import type { MessageChunk } from '@archon/providers/types';
 import { createLogger } from '@archon/paths';
 import { MessagePersistence } from './web/persistence';
 import { SSETransport, type SSEWriter } from './web/transport';
